// src/app/data-access/api/model/responsePageTransactions.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PageTransactions } from './pageTransactions';


export interface ResponsePageTransactions { 
    code?: number;
    message?: string;
    data?: PageTransactions;
}